requirements.txt
beautifulsoup4
python-dotenv
selenium
webdriver-manager
supabase
pytz
